chore(server): move connect handler from task/ to identity/
The Connect RPC handles the WebSocket auth handshake — it validates
`agentKey` / `sessionToken` and resolves the authenticated context.
That's an identity concern; the handler belongs in identity/.

- `git mv` `task/handlers/connect.handlers.ts` → `identity/handlers/`.
- Fix the `../services/conversation.service.js` import to
  `../../task/services/...` after the cross-folder move.
- `app/server.ts`: update import path.
- `transport/layer-tags.ts`: update doc citation to new path.
