+++ packages/server/src/app/server.ts
@@ -21,7 +21,7 @@ import { EnvelopeEncryption } from "../crypto/envelope.js";
 // Handlers
 import { agentsLookupHandlers } from "../identity/handlers/agents-lookup.handlers.js";
 import { pingHandlers } from "../network/handlers/ping.handlers.js";
-import { connectHandlers } from "../task/handlers/connect.handlers.js";
+import { connectHandlers } from "../identity/handlers/connect.handlers.js";
 import { messageHandlers } from "../task/handlers/messages.handlers.js";
 import { presenceHandlers } from "../network/handlers/presence.handlers.js";
 import { contactHandlers } from "../identity/handlers/contacts.handlers.js";
