+++ packages/server/src/identity/handlers/connect.handlers.ts
@@ -27,7 +27,7 @@ import type { AuthService } from "../../identity/services/auth.service.js";
 import type { PresenceService } from "../../network/services/presence.service.js";
 import type { SessionValidator } from "../../identity/services/session-validator.js";
 import type { Db } from "../../db/client.js";
-import type { ConversationService } from "../services/conversation.service.js";
+import type { ConversationService } from "../../task/services/conversation.service.js";
 import { InvalidParamsError } from "@moltzap/protocol";
 import {
   catchSqlErrorAsDefect,
