+++ packages/server/src/transport/layer-tags.ts
@@ -98,12 +98,12 @@ export type NetworkTags =
   | PresenceServiceTag;
 
 /**
- * Task-layer allowlist: conversations, messages, tasks, contacts.
+ * Task-layer allowlist: conversations, messages, tasks.
  * Includes `LeaseRegistryTag` (admission gate for `messages/send`
  * dispatch leases — yielded by `messages.handlers.ts`) and
- * `SessionValidatorTag` (yielded by the Connect handler post-split,
- * which lives at `task/handlers/connect.handlers.ts`). The Connect
- * handler is task-tier because its body pulls cross-cutting services
+ * `SessionValidatorTag` (yielded by the Connect handler at
+ * `identity/handlers/connect.handlers.ts`). The Connect handler runs
+ * at task-tier because its body pulls cross-cutting services
  * spanning network (connections, presence) AND task (conversation
  * resolution for presence fan-out).
  */
